+++ src/components/UpgradePromptDialog.jsx
@@ -36,7 +36,7 @@ export function UpgradePromptDialog({ open, onOpenChange }) {
         </AlertDialogHeader>
         <AlertDialogFooter>
           <AlertDialogCancel>Maybe Later</AlertDialogCancel>
-          <AlertDialogAction onClick={handleUpgrade}>Upgrade to Pro</AlertDialogAction>
+          <AlertDialogAction onClick={handleUpgrade}>Upgrade</AlertDialogAction>
         </AlertDialogFooter>
       </AlertDialogContent>
     </AlertDialog>